+++ src/cpu/sphere.h
@@ -7,7 +7,7 @@
 
 class sphere : public hittable {
     public:
-        sphere(const point3& center, double radius, shared_ptr<material> mat) : center(center), radius(std::fmax(0, radius)) , mat(mat) {}
+        sphere(const point3& center, float radius, shared_ptr<material> mat) : center(center), radius(std::fmax(0, radius)) , mat(mat) {}
 
 
         bool hit(const ray& r, interval ray_t, hit_record& rec) const override{
@@ -44,7 +44,7 @@ class sphere : public hittable {
         }
     private:
         point3 center;
-        double radius;
+        float radius;
         shared_ptr<material> mat;
 
 };
